requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai>=1.0.0
azure-cosmos